+++ attributes/php_driver.rb
@@ -8,7 +8,7 @@
 #
 
 # Latest build numbers for each CUBRID PHP version in the form of 'version'=>'build_number'.
-build_numbers = {'9.1.0' => '0003', '9.0.0' => '0001', '8.4.3' => '0001', '8.4.1' => '0006'}
+build_numbers = {'9.1.0' => '0004', '9.0.0' => '0001', '8.4.3' => '0001', '8.4.1' => '0006'}
 
 # The default version of CUBRID PHP driver to install.
 default['cubrid']['version'] = "9.1.0"